// src/components/ReferanslarSayfasiKlinikVaka/types.ts
// This file is auto-generated — do not edit manually.
import type { IkasImage } from "@ikas/bp-storefront";

export interface Props {
  sectionAnchorId?: string;
  caseImageAlt?: string;
  caseDescription?: string;
  caseImageUrl?: IkasImage | null;
  caseTitle?: string;
  storyPoint1Text?: string;
  caseKicker?: string;
  caseCaptionText?: string;
  caseCaptionTitle?: string;
  storyPoint3Text?: string;
  storyPoint4Text?: string;
  storyPoint5Text?: string;
  storyPoint2Text?: string;
  storyPoint2TextEn?: string;
  storyPoint3TextEn?: string;
  storyPoint4TextEn?: string;
  caseCaptionTextEn?: string;
  storyPoint1TextEn?: string;
  caseCaptionTitleEn?: string;
  caseImageAltEn?: string;
  caseDescriptionEn?: string;
  caseTitleEn?: string;
  caseKickerEn?: string;
  storyPoint5TextEn?: string;
  backgroundColor?: string;
  paddingBottom?: number;
  paddingTop?: number;
  maxWidth?: number;
  lineColor?: string;
  textColor?: string;
  mutedTextColor?: string;
  showStoryPoint4?: boolean;
  showStoryPoint2?: boolean;
  showStoryPoint1?: boolean;
  showCaseStudy?: boolean;
  showStoryPoint3?: boolean;
  showStoryPoint5?: boolean;
}